remove manual Guacamole setup from bootstrap.tpl
I don't like the idea of setting up Guacamole via the user data script
using bash commands. It's too brittle. Going to try and figure out if it
can be done using containers, which would be much easier to manage.

Need to figure out if there's an ARM container for Guacamole, guacd and
MySQL